+++ GVRf/Framework/framework/src/main/jni/objects/uniform_block.cpp
@@ -17,7 +17,7 @@
 #include <cctype>
 #include <sstream>
 #include "util/gvr_gl.h"
-#include "vulkan/vulkanCore.h"
+
 namespace gvr {
 
     UniformBlock::UniformBlock(const std::string& descriptor) :
@@ -428,241 +428,4 @@ namespace gvr {
         return os.str();
     }
 
-    GLUniformBlock::GLUniformBlock(const std::string& descriptor) :
-            UniformBlock(descriptor),
-            GLBlockIndex(-1),
-            GLBindingPoint(-1),
-            GLOffset(0),
-            GLBuffer(0)
-    {
-    }
-
-    GLUniformBlock::GLUniformBlock() :
-            UniformBlock(),
-            GLBlockIndex(-1),
-            GLOffset(0),
-            GLBuffer(0),
-            GLBindingPoint(-1)
-    {
-
-    }
-
-    void GLUniformBlock::bindBuffer(GLuint programId)
-    {
-
-        if (GLBindingPoint < 0)
-            return;
-        if (GLBlockIndex < 0)
-        {
-            GLBlockIndex = glGetUniformBlockIndex(programId, BlockName.c_str());
-            if (GLBlockIndex < 0)
-            {
-                LOGE("UniformBlock: ERROR: cannot find block named %s\n", BlockName.c_str());
-                return;
-            }
-
-            glGenBuffers(1, &GLBuffer);
-            glBindBuffer(GL_UNIFORM_BUFFER, GLBuffer);
-            glBufferData(GL_UNIFORM_BUFFER, TotalSize, NULL, GL_DYNAMIC_DRAW);
-            glUniformBlockBinding(programId, GLBlockIndex, GLBindingPoint);
-            glBindBufferBase(GL_UNIFORM_BUFFER, GLBindingPoint, GLBuffer);
-            checkGlError("bindUBO ");
-            LOGV("SHADER: UniformBlock: %s bound to #%d at index %d buffer = %d\n", BlockName.c_str(), GLBindingPoint, GLBlockIndex, GLBuffer);
-        }
-        else {
-            glBindBuffer(GL_UNIFORM_BUFFER, GLBuffer);
-            glBindBufferBase(GL_UNIFORM_BUFFER, GLBindingPoint, GLBuffer);
-        }
-    }
-
-    void GLUniformBlock::render(GLuint programId)
-    {
-        auto it = Dirty.find(programId);
-
-        if (it != Dirty.end() && !it->second)
-            return;
-
-        //  LOGE("it should not come hrere");
-        Dirty[programId] = false;
-        if (GLBuffer == 0 )
-            bindBuffer(programId);
-        if (GLBuffer >= 0)
-        {
-            glBindBuffer(GL_UNIFORM_BUFFER, GLBuffer);
-            glBindBufferBase(GL_UNIFORM_BUFFER, GLBindingPoint, GLBuffer);
-            glBufferSubData(GL_UNIFORM_BUFFER, GLOffset, TotalSize, UniformData);
-            LOGV("SHADER: UniformBlock: offset %d : total Size %d\n", GLOffset, TotalSize);
-        }
-    }
-
-    void GLUniformBlock::dump(GLuint programID, int blockIndex)
-    {
-        // get size of name of the uniform block
-        GLint nameLength;
-        glGetActiveUniformBlockiv(programID, blockIndex, GL_UNIFORM_BLOCK_NAME_LENGTH, &nameLength);
-
-        // get name of uniform block
-        GLchar blockName[nameLength];
-        glGetActiveUniformBlockName(programID, blockIndex, nameLength, NULL, blockName);
-
-        // get size of uniform block in bytes
-        GLint byteSize;
-        glGetActiveUniformBlockiv(programID, blockIndex, GL_UNIFORM_BLOCK_DATA_SIZE, &byteSize);
-
-        // get number of uniform variables in uniform block
-        GLint numberOfUniformsInBlock;
-        glGetActiveUniformBlockiv(programID, blockIndex,
-                                  GL_UNIFORM_BLOCK_ACTIVE_UNIFORMS, &numberOfUniformsInBlock);
-
-        // get indices of uniform variables in uniform block
-        GLint uniformsIndices[numberOfUniformsInBlock];
-        glGetActiveUniformBlockiv(programID, blockIndex, GL_UNIFORM_BLOCK_ACTIVE_UNIFORM_INDICES, uniformsIndices);
-        LOGV("UniformBlock: %s %d bytes\n", blockName, byteSize);
-
-        // get parameters of all uniform variables in uniform block
-        for (int uniformMember=0; uniformMember<numberOfUniformsInBlock; uniformMember++)
-        {
-            if (uniformsIndices[uniformMember] > 0)
-            {
-                // index of uniform variable
-                GLuint tUniformIndex = uniformsIndices[uniformMember];
-
-                uniformsIndices[uniformMember];
-                GLint uniformNameLength, uniformOffset, uniformSize;
-                GLint uniformType, arrayStride, matrixStride;
-
-                // get length of name of uniform variable
-                glGetActiveUniformsiv(programID, 1, &tUniformIndex,
-                                      GL_UNIFORM_NAME_LENGTH, &uniformNameLength);
-                // get name of uniform variable
-                GLchar uniformName[uniformNameLength];
-                glGetActiveUniform(programID, tUniformIndex, uniformNameLength,
-                                   NULL, NULL, NULL, uniformName);
-
-                // get offset of uniform variable related to start of uniform block
-                glGetActiveUniformsiv(programID, 1, &tUniformIndex,
-                                      GL_UNIFORM_OFFSET, &uniformOffset);
-                // get size of uniform variable (number of elements)
-                glGetActiveUniformsiv(programID, 1, &tUniformIndex,
-                                      GL_UNIFORM_SIZE, &uniformSize);
-                // get type of uniform variable (size depends on this value)
-                glGetActiveUniformsiv(programID, 1, &tUniformIndex,
-                                      GL_UNIFORM_TYPE, &uniformType);
-                // offset between two elements of the array
-                glGetActiveUniformsiv(programID, 1, &tUniformIndex,
-                                      GL_UNIFORM_ARRAY_STRIDE, &arrayStride);
-                // offset between two vectors in matrix
-                glGetActiveUniformsiv(programID, 1, &tUniformIndex,
-                                      GL_UNIFORM_MATRIX_STRIDE, &matrixStride);
-
-                // Size of uniform variable in bytes
-                byteSize = uniformSize * sizeFromUniformType(uniformType);
-                LOGV("UniformBlock: %s GL offset = %d, byteSize = %d\n", uniformName, uniformOffset, byteSize);
-            }
-        }
-    }
-
-    GLint GLUniformBlock::sizeFromUniformType(GLint type)
-    {
-        GLint s;
-
-#define UNI_CASE(type, numElementsInType, elementType) \
-       case type : s = numElementsInType * sizeof(elementType); break;
-
-        switch (type)
-        {
-            UNI_CASE(GL_FLOAT, 1, GLfloat);
-            UNI_CASE(GL_FLOAT_VEC2, 2, GLfloat);
-            UNI_CASE(GL_FLOAT_VEC3, 3, GLfloat);
-            UNI_CASE(GL_FLOAT_VEC4, 4, GLfloat);
-            UNI_CASE(GL_INT, 1, GLint);
-            UNI_CASE(GL_INT_VEC2, 2, GLint);
-            UNI_CASE(GL_INT_VEC3, 3, GLint);
-            UNI_CASE(GL_INT_VEC4, 4, GLint);
-            UNI_CASE(GL_UNSIGNED_INT, 1, GLuint);
-            UNI_CASE(GL_UNSIGNED_INT_VEC2, 2, GLuint);
-            UNI_CASE(GL_UNSIGNED_INT_VEC3, 3, GLuint);
-            UNI_CASE(GL_UNSIGNED_INT_VEC4, 4, GLuint);
-            UNI_CASE(GL_BOOL, 1, GLboolean);
-            UNI_CASE(GL_BOOL_VEC2, 2, GLboolean);
-            UNI_CASE(GL_BOOL_VEC3, 3, GLboolean);
-            UNI_CASE(GL_BOOL_VEC4, 4, GLboolean);
-            UNI_CASE(GL_FLOAT_MAT2, 4, GLfloat);
-            UNI_CASE(GL_FLOAT_MAT3, 9, GLfloat);
-            UNI_CASE(GL_FLOAT_MAT4, 16, GLfloat);
-            UNI_CASE(GL_FLOAT_MAT2x3, 6, GLfloat);
-            UNI_CASE(GL_FLOAT_MAT2x4, 8, GLfloat);
-            UNI_CASE(GL_FLOAT_MAT3x2, 6, GLfloat);
-            UNI_CASE(GL_FLOAT_MAT3x4, 12, GLfloat);
-            UNI_CASE(GL_FLOAT_MAT4x2, 8, GLfloat);
-            UNI_CASE(GL_FLOAT_MAT4x3, 12, GLfloat);
-            default : s = 0; break;
-        }
-        return s;
-    }
-
-    VulkanUniformBlock::VulkanUniformBlock(const std::string& descriptor) :
-            UniformBlock(descriptor), buffer_init_(false){
-
-        LOGE("pararameter");
-    }
-
-    VulkanUniformBlock::VulkanUniformBlock() :
-            UniformBlock(), buffer_init_(false)
-    {
-        LOGE("default");
-    }
-    void VulkanUniformBlock::updateBuffer(VkDevice &device,VulkanCore* vk){
-
-        //  if(!buffer_init_)
-        //      createBuffer(device, vk);
-
-        VkResult ret = VK_SUCCESS;
-        uint8_t *pData;
-
-        ret = vkMapMemory(device, m_bufferInfo.mem, 0, m_bufferInfo.allocSize, 0, (void **) &pData);
-        assert(!ret);
-
-        memcpy(pData, UniformData, TotalSize);
-
-        vkUnmapMemory(device, m_bufferInfo.mem);
-
-    }
-    void VulkanUniformBlock::createBuffer(VkDevice &device,VulkanCore* vk){
-
-        VkResult err = VK_SUCCESS;
-        memset(&m_bufferInfo, 0, sizeof(m_bufferInfo));
-        //err = vkCreateBuffer(m_device, &bufferCreateInfo, NULL, &m_modelViewMatrixUniform.buf);
-        err = vkCreateBuffer(device, gvr::BufferCreateInfo(TotalSize, VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT), NULL, &m_bufferInfo.buf);
-        assert(!err);
-        // Obtain the requirements on memory for this buffer
-        VkMemoryRequirements mem_reqs;
-        vkGetBufferMemoryRequirements(device, m_bufferInfo.buf, &mem_reqs);
-        assert(!err);
-
-        // And allocate memory according to those requirements
-        VkMemoryAllocateInfo memoryAllocateInfo;
-        memoryAllocateInfo.sType = VK_STRUCTURE_TYPE_MEMORY_ALLOCATE_INFO;
-        memoryAllocateInfo.pNext = NULL;
-        memoryAllocateInfo.allocationSize = 0;
-        memoryAllocateInfo.memoryTypeIndex = 0;
-        memoryAllocateInfo.allocationSize  = mem_reqs.size;
-        bool pass = vk->GetMemoryTypeFromProperties(mem_reqs.memoryTypeBits, VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT, &memoryAllocateInfo.memoryTypeIndex);
-        assert(pass);
-
-        // We keep the size of the allocation for remapping it later when we update contents
-        m_bufferInfo.allocSize = memoryAllocateInfo.allocationSize;
-
-        err = vkAllocateMemory(device, &memoryAllocateInfo, NULL, &m_bufferInfo.mem);
-        assert(!err);
-
-        // Bind our buffer to the memory
-        err = vkBindBufferMemory(device, m_bufferInfo.buf, m_bufferInfo.mem, 0);
-        assert(!err);
-
-        m_bufferInfo.bufferInfo.buffer = m_bufferInfo.buf;
-        m_bufferInfo.bufferInfo.offset = 0;
-        m_bufferInfo.bufferInfo.range = TotalSize;
-
-    }
 }
